requirements.txt
langchain
openai
faiss-cpu
sentence-transformers